avoid unnecessary copy of 'GraphNode'
When compiling the command line binary with GCC 11.4.0, the compiler observes:

  source/GraphDrawing/graph.cpp: In member function ‘bool Graph::addNode(GraphNode)’:
  source/GraphDrawing/graph.cpp:93:19: warning: implicitly-declared ‘GraphNode::GraphNode(const GraphNode&)’ is deprecated [-Wdeprecated-copy]
    93 |   if (containsNode(node))
       |       ~~~~~~~~~~~~^~~~~~
  In file included from source/GraphDrawing/graph.h:9,
                   from source/GraphDrawing/graph.cpp:1:
  source/GraphDrawing/graphnode.h:51:14: note: because ‘GraphNode’ has user-provided ‘GraphNode& GraphNode::operator=(const GraphNode&)’
    51 |   GraphNode &operator=(const GraphNode &);
       |              ^~~~~~~~

What is happening here is that passing `node` by value causes a call to the
(implicitly defined) `GraphNode` copy constructor. The compiler claims the copy
constructor is deprecated because `GraphNode` has a user-defined copy assignment
operator and it is expecting the class to therefore also implement the copy
constructor.¹ We could (and probably in future should) implement the copy
constructor. But this change instead avoids the copy altogether, which should be
a minor performance win.

¹ https://en.cppreference.com/w/cpp/language/rule_of_three