remove unnecessary explicit 'GraphNode' copy assignment operator
`GraphNode`’s data members all have POD or STL class types. So the default copy
assignment operator does the right thing. The only difference between it and the
explicitly implemented one was the branch to shortcut assigning a `GraphNode` to
itself. But nothing in the code base appears to rely on this (no `a = (b = c)`
style code with `GraphNode`s).

Removing this squashes 6 -Wdeprecated-copy warnings when compiling on Linux with
GCC 11.4.0. This is because the compiler no longer believes the copy constructor
to be deprecated because it itself is implementing all four of copy/move
constructor/assignment operator.