reduce casting and improve type safety of 'Polynomial' cloning
C++ supports so-called “covariant return types”¹ that permit an overriding
function to return a “narrower” type than its parent. This allows implementing
a copying helper that (1) returns a more precise type and (2) does not require
casting at the call site, thus removing an opportunity for mistakes. This change
allows the compiler to validate calls to `Polynomial::clone` and its children.

¹ https://en.wikipedia.org/wiki/Covariant_return_type