reduce casting and improve type safety of 'Term' cloning
This is the equivalent of 64ca5a0d0644cd7816bafa88b52ea10e8f7816d6, but applied
to the `Term` hierarchy.

C++ supports so-called “covariant return types”¹ that permit an overriding
function to return a “narrower” type than its parent. This allows implementing
a copying helper that (1) returns a more precise type and (2) does not require
casting at the call site, thus removing an opportunity for mistakes. This change
allows the compiler to validate calls to `Polynomial::clone` and its children.

¹ https://en.wikipedia.org/wiki/Covariant_return_type