remove use of 'alloca'
The `alloca` function allocates memory on the stack. If the requested size is
too large, a silent stack overflow occurs and the program will crash or
misbehave. In this case where the allocated size is not known statically, it is
safer to use `std::string`. Small String Optimization¹ also means short command
names will still be stack-allocated when possible.

¹ https://pvs-studio.com/en/blog/terms/6658/