incorporate Qt GUI compilation into CMake build system
This is part of a set of steps to unify build system files.

This change attempts to replicate the semantics of source/gGCLC.pro in CMake.
Unlike the prior porting of Makefile and Makefile.win, the old build file is not
removed. It is assumed users may still wish to use a Qt Creator-driven workflow.

Notable deviations from source/gGCLC.pro:

  1. Whether the GUI is built in the CMake build system is controlled by the
     `-Dgui=…` option. If this is omitted, it defaults to `AUTO` which builds
     the GUI iff Qt 6 is installed.

  2. The `QT_DISABLE_DEPRECATED_UP_TO=0x050F00` define has been dropped. It
     seems this is mainly useful when porting from Qt 5 to Qt 6, which has
     already been completed.

  3. The `-fno-sized-deallocation` CXXFLAG has been dropped. It is not clear to
     me why this flag was being used.

  4. The `-static` option used in CI on Windows has been dropped. This
     reportedly was not working correctly. Investigating how to re-enable it
     will be part of future work.

Github: #23