add 'install' targets to the build system
These are a bit non-standard compared to the usual CMake pattern for this.
Typically one would `include(GNUInstallDirs)` and then use the variable
`$CMAKE_INSTALL_BINDIR` as the destination for executables. However the current
desire is to have them at the top level, as this will not be their final
destination.