port Makefile.web to CMake
This deviates in some ways from the logic in Makefile.web:

  1. Patching of the generated JavaScript is done into a separate file instead
     of in-place.

  2. The first part of the patching is done with `echo` instead of `sed` for
     simplicity.

  3. Generated files are copied from executable/ into their source directory
     locations instead of being moved.

  4. `-O2` is no longer hard coded as a compiler optimisation level.

  5. The C++ standard is explicitly set to C++17.

  6. emcc/emc++ options are all passed as `-s<name>=<value>` instead of
     `-s <name>=<value>`. The latter causes CMake to believe `-s` is a redundant
     repeated option and can be de-duped.

Rather than integrate this into the existing CMake project, this change creates
a new CMake project. It seems the recommended way to use Emscripten in
combination with CMake is via the `emcmake` wrapper which sets a specific
toolchain for the entire compilation.

Github: closes #23