reference count 'Term' hierarchy with 'std::shared_ptr'
This avoids the need for manual `Dispose` calls, removing an opportunity for
accidental omissions. By leaning on the standard library, we also get a
reference counting implementation optimised for the target platform.

This is significantly more complex and cross-cutting than the previous
`std::shared_ptr` migrations. Unfortunately there seems to be no easy way to
break this into smaller steps.

A suboptimal part of this change is that the `Clone` methods can no longer use
covariant return types for type safety:

  -virtual Term *Term::Clone() = 0;
  -UTerm *UTerm::Clone() override;
  -XTerm *XTerm::Clone() override;
  +virtual std::shared_ptr<Term> Term::Clone() = 0;
  +std::shared_ptr<Term> UTerm::Clone() override;
  +std::shared_ptr<Term> XTerm::Clone() override;

Note in the above how all `Clone` methods now need to return a pointer to the
base class. This means call sites that need a pointer to the derived class now
need to use a dynamic cast:

  -XTerm *copy = my_xterm->Clone();
  +std::shared_ptr<XTerm> copy =
  +  std::dynamic_pointer_cast<XTerm>(my_xterm->Clone());

The main weakness in the new code is that type checking is now postponed to
runtime; the compiler can no longer catch mistakes like cloning an `XTerm` into
a `UTerm` pointer.