manage constructed lines with 'std::unique_ptr'
This avoids a problem wherein exceptions can leak memory:

  1: Line *l = new Line(…);
  2: my_vec.emplace_back(l);

If allocation in (1) succeeds but then `my_vec` needs expansion to store `l` and
the expansion in (2) fails, an exception will be thrown and the memory allocated
in (1) will be leaked. Using a `std::unique_ptr` ensures this memory gets
deallocated on exception.