regain covariance in '*Term::Clone()'
The commit message for a31a0a0cecbc07d2f90a70d43d4e0b5f98a6a568 discusses a
weakness wherein some type safety checks that could previously happen at
compile-time were degraded to happening at runtime. The present change uses a
technique¹ to regain these compile-time checks.

Prior to this change, it was possible to write code like the following:

  std::shared_ptr<XTerm> xterm = std::make_shared<XTerm>();
  std::shared_ptr<UTerm> uterm =
    std::dynamic_pointer_cast<UTerm>(xterm->Clone());

This clones an `XTerm` into a `UTerm`, something that is disallowed. However
this would only be discovered through the `dynamic_pointer_cast` returning
`nullptr` at runtime. That is, it is not possible for the compiler to statically
observe the type safety violation here.

Prior to a31a0a0cecbc07d2f90a70d43d4e0b5f98a6a568, it would not have been
possible to write such code:

  XTerm *xterm = new XTerm();
  UTerm *uterm = xterm->Clone(); // ← fails to compile

With the present change, we re-establish this property:

  std::shared_ptr<XTerm> xterm = std::make_shared<XTerm>();
  std::shared_ptr<UTerm> uterm = xterm->Clone(); // ← fails to compile

This change does not go as far as using CRTP² to abbreviate the boiler plate, as
discussed in the linked blog post. For a shallow type hierarchy like `Term`, the
increased complexity for the reader of this code does not seem worth it.

¹ https://www.fluentcpp.com/2017/09/12/how-to-return-a-smart-pointer-and-use-covariance/
² https://en.wikipedia.org/wiki/Curiously_recurring_template_pattern