Reduce::ReduceLineCircleIntersection: fix memory leak of 'xp22'
Due to the `#if 1` block taking an early exit, the cleanup `.Dispose` call was
never reached. This change takes advantage of the scope of this variable to
observe that heap allocation is not required. Note that the leak here is latent
as the initial `return false` ensures most of this function is never executed.